Getting stuff to work good enough doesn't mean the code is good. My prior modifications were actually ugly as hell, and I fixed them now.
Compared to the normal artwork rules (assuming those work properly, which I can't really figure out since I don't use the majority of those), the only difference is that a file named 'cover' gets brownie points, being chosen over a file with 'front' in the name. A file that merely has 'cover' in the name is still ranked lower.

Also, I believe there wa a bug regarding files named after their albums, and I hope to have fixed those. If they weren't bugged in the first place, feel free to inform me how the code works and I'll revert it, but I think this is how it was supposed to be.